% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_arborea.R
\name{create_arborea}
\alias{create_arborea}
\title{Création de la structure du projet}
\usage{
create_arborea(path)
}
\arguments{
\item{path}{chemin d'accès du projet R}
}
\value{
les fichiers et dossiers
}
\description{
Fonction qui créer l'arborescence des dossiers, ainsi que les dossiers de config
}
\examples{
\dontrun{
create_arborea("my_project_path")
}
}
